Add Megatron true-on-policy namespace
