+++ megatron/core/true_on_policy/__init__.py
@@ -1,0 +1,2 @@
+"""True-on-policy backend components for Megatron Core."""
+
